move /api/reset api route to settings.py
